Footer and menu are now loaded seperately from main content by the browser using Javascript.  Common for static sites.
This allows search engines to reindex only when the content on a particular page
changes;  footer changes, or changes in the menu (say from adding anothe page) won't
result in a change to the page as loaded by the user-agent.

So when there is a change, the search engine reindexing budget for a doclog site won't be spent on
unchnaged content.